requirements.txt
google-api-python-client
pytube
google-cloud-speech
spacy
